Move Level into the crate root
