+++ askama_parser/src/expr.rs
@@ -10,7 +10,7 @@ use nom::sequence::{pair, preceded, terminated, tuple};
 use nom::{error_position, IResult};
 
 use super::{
-    char_lit, identifier, not_ws, num_lit, path_or_identifier, str_lit, ws, PathOrIdentifier,
+    char_lit, identifier, not_ws, num_lit, path_or_identifier, str_lit, ws, Level, PathOrIdentifier,
 };
 
 macro_rules! expr_prec_layer {
@@ -362,19 +362,4 @@ impl<'a> Suffix<'a> {
     fn r#try(i: &'a str) -> IResult<&'a str, Self> {
         map(preceded(take_till(not_ws), char('?')), |_| Self::Try)(i)
     }
-}
-
-#[derive(Clone, Copy, Default)]
-pub(crate) struct Level(u8);
-
-impl Level {
-    fn nest(self, i: &str) -> Result<Level, nom::Err<nom::error::Error<&str>>> {
-        if self.0 >= Self::MAX_EXPR_DEPTH {
-            return Err(nom::Err::Failure(error_position!(i, ErrorKind::TooLarge)));
-        }
-
-        Ok(Level(self.0 + 1))
-    }
-
-    const MAX_EXPR_DEPTH: u8 = 64;
 }
